These aren't acronyms, so use regular case to enable proper loading
These aren't widely known acronyms, so Zeitwerk and rails need to be told about
them.  We should avoid using acronyms and default to using the "rails"
convention unless it's warranted to specify acronyms.  All other usages of the
same acronyms would need to be enforced in locations so it's not without side
effects.

It's just not worth it here.