Use the rails 7.1+ serialize interface
Type was changed from positional to a keyword argument 'type' in 7.1.
The positional argument is removed in rails 7.2.

We have a helper bridge module prepending support for either kwargs
or positional arguments passed to serialize which converts to positional
for rails 7.0 and kwargs for 7.1+.  This was added and automatically
loaded in the rails app via the engine in schema:
https://github.com/ManageIQ/manageiq-schema/pull/756

Similar change as core: https://github.com/ManageIQ/manageiq/pull/23253